remove half precision

+++ src/trainer.py
@@ -33,8 +33,7 @@ def regular_train(epochs: int = 10, batch_size: int = 50, lr: float = 1e-4):
         collate_fn=collate_fn,
     )
 
-    # Model
-    model = Decoder().half().to(DEVICE)
+    model = Decoder().to(DEVICE)
 
     # Optimizer
     optimizer = torch.optim.Adam(model.parameters(), lr=lr)
@@ -126,8 +125,8 @@ def distillation_train(epochs: int = 10, batch_size: int = 50, lr: float = 1e-4)
         collate_fn=collate_fn,
     )
 
-    student = Decoder().half().to(DEVICE)
-    teacher = MODEL.half().to(DEVICE)
+    student = Decoder().to(DEVICE)
+    teacher = MODEL.to(DEVICE)
 
     optimizer = torch.optim.Adam(student.parameters(), lr=lr)
